data/documents/requirements.txt
lxml
newspaper3k
lxml_html_clean
